+++ openvino_xai/methods/__init__.py
@@ -3,6 +3,7 @@
 """
 XAI algorithms.
 """
+from openvino_xai.methods.black_box.aise import AISE
 from openvino_xai.methods.black_box.rise import RISE
 from openvino_xai.methods.white_box.activation_map import ActivationMap
 from openvino_xai.methods.white_box.base import WhiteBoxMethod
@@ -23,4 +24,5 @@ __all__ = [
     "ViTReciproCAM",
     "DetClassProbabilityMap",
     "RISE",
+    "AISE",
 ]