+++ openvino_xai/methods/black_box/rise.py
@@ -1,25 +1,27 @@
 # Copyright (C) 2023-2024 Intel Corporation
 # SPDX-License-Identifier: Apache-2.0
 
-from typing import Callable, List, Mapping, Tuple
+from typing import Callable, Dict, List, Mapping, Tuple
 
 import cv2
 import numpy as np
 import openvino.runtime as ov
 from tqdm import tqdm
 
-from openvino_xai.common.utils import IdentityPreprocessFN, scaling
-from openvino_xai.methods.black_box.base import BlackBoxXAIMethod
+from openvino_xai.common.utils import IdentityPreprocessFN, is_bhwc_layout, scaling
+from openvino_xai.methods.black_box.base import BlackBoxXAIMethod, Preset
 
 
 class RISE(BlackBoxXAIMethod):
-    """RISE explains classification models in black-box mode using RISE (https://arxiv.org/abs/1806.07421).
+    """RISE explains classification models in black-box mode using
+    'RISE: Randomized Input Sampling for Explanation of Black-box Models' paper
+    (https://arxiv.org/abs/1806.07421).
 
     :param model: OpenVINO model.
     :type model: ov.Model
-    :param postprocess_fn: Preprocessing function that extract scores from IR model output.
-    :type postprocess_fn: Callable[[Mapping], np.ndarray]
-    :param preprocess_fn: Preprocessing function, identity function by default
+    :param postprocess_fn: Post-processing function that extract scores from IR model output.
+    :type postprocess_fn: Callable[[OVDict], np.ndarray]
+    :param preprocess_fn: Pre-processing function, identity function by default
         (assume input images are already preprocessed by user).
     :type preprocess_fn: Callable[[np.ndarray], np.ndarray]
     :param device_name: Device type name.
@@ -42,28 +44,26 @@ class RISE(BlackBoxXAIMethod):
         if prepare_model:
             self.prepare_model()
 
-    def prepare_model(self, load_model: bool = True) -> ov.Model:
-        if load_model:
-            self.load_model()
-        return self._model
-
     def generate_saliency_map(
         self,
         data: np.ndarray,
-        explain_target_indices: List[int] | None = None,
-        num_masks: int = 5000,
+        target_indices: List[int] | None = None,
+        preset: Preset = Preset.BALANCE,
+        num_masks: int | None = None,
         num_cells: int = 8,
         prob: float = 0.5,
         seed: int = 0,
         scale_output: bool = True,
-    ):
+    ) -> np.ndarray | Dict[int, np.ndarray]:
         """
         Generates inference result of the RISE algorithm.
 
         :param data: Input image.
         :type data: np.ndarray
-        :param explain_target_indices: List of target indices to explain.
-        :type explain_target_indices: List[int]
+        :param target_indices: List of target indices to explain.
+        :type target_indices: List[int]
+        :param preset: Speed-Quality preset, defines predefined configurations that manage speed-quality tradeoff.
+        :type preset: Preset
         :param num_masks: Number of generated masks to aggregate.
         :type num_masks: int
         :param num_cells: Number of cells for low-dimensional RISE
@@ -79,19 +79,44 @@ class RISE(BlackBoxXAIMethod):
         """
         data_preprocessed = self.preprocess_fn(data)
 
+        num_masks = self._preset_parameters(preset, num_masks)
+
         saliency_maps = self._run_synchronous_explanation(
             data_preprocessed,
-            explain_target_indices,
+            target_indices,
             num_masks,
             num_cells,
             prob,
             seed,
         )
 
-        if scale_output:
-            saliency_maps = scaling(saliency_maps)
-        saliency_maps = np.expand_dims(saliency_maps, axis=0)
+        if isinstance(saliency_maps, np.ndarray):
+            if scale_output:
+                saliency_maps = scaling(saliency_maps)
+            saliency_maps = np.expand_dims(saliency_maps, axis=0)
+        elif isinstance(saliency_maps, dict):
+            for target in saliency_maps:
+                if scale_output:
+                    saliency_maps[target] = scaling(saliency_maps[target])
         return saliency_maps
+
+    @staticmethod
+    def _preset_parameters(
+        preset: Preset,
+        num_masks: int | None = None,
+    ) -> int:
+        # TODO (negvet): preset num_cells
+        if num_masks is not None:
+            return num_masks
+
+        if preset == Preset.SPEED:
+            return 2000
+        elif preset == Preset.BALANCE:
+            return 5000
+        elif preset == Preset.QUALITY:
+            return 8000
+        else:
+            raise ValueError(f"Preset {preset} is not supported.")
 
     def _run_synchronous_explanation(
         self,
@@ -102,13 +127,9 @@ class RISE(BlackBoxXAIMethod):
         prob: float,
         seed: int,
     ) -> np.ndarray:
-        from openvino_xai.explainer.utils import is_bhwc_layout
-
         input_size = data_preprocessed.shape[1:3] if is_bhwc_layout(data_preprocessed) else data_preprocessed.shape[2:4]
 
-        forward_output = self.model_forward(data_preprocessed, preprocess=False)
-        logits = self.postprocess_fn(forward_output)
-        _, num_classes = logits.shape
+        num_classes = self.get_num_classes(data_preprocessed)
 
         if target_classes is None:
             num_targets = num_classes
@@ -117,7 +138,7 @@ class RISE(BlackBoxXAIMethod):
 
         rand_generator = np.random.default_rng(seed=seed)
 
-        sal_maps = np.zeros((num_targets, input_size[0], input_size[1]))
+        saliency_maps = np.zeros((num_targets, input_size[0], input_size[1]))
         for _ in tqdm(range(0, num_masks), desc="Explaining in synchronous mode"):
             mask = self._generate_mask(input_size, num_cells, prob, rand_generator)
             # Add channel dimensions for masks
@@ -130,11 +151,11 @@ class RISE(BlackBoxXAIMethod):
             raw_scores = self.postprocess_fn(forward_output)
 
             sal = self._get_scored_mask(raw_scores, mask, target_classes)
-            sal_maps += sal
+            saliency_maps += sal
 
         if target_classes is not None:
-            sal_maps = self._reconstruct_sparce_saliency_map(sal_maps, num_classes, input_size, target_classes)
-        return sal_maps
+            saliency_maps = self._reformat_as_dict(saliency_maps, target_classes)
+        return saliency_maps
 
     @staticmethod
     def _get_scored_mask(raw_scores: np.ndarray, mask: np.ndarray, target_classes: List[int] | None) -> np.ndarray:
@@ -144,15 +165,11 @@ class RISE(BlackBoxXAIMethod):
             return raw_scores.reshape(-1, 1, 1) * mask
 
     @staticmethod
-    def _reconstruct_sparce_saliency_map(
-        sal_maps: np.ndarray, num_classes: int, input_size, target_classes: List[int] | None
+    def _reformat_as_dict(
+        saliency_maps: np.ndarray,
+        target_classes: List[int] | None,
     ) -> np.ndarray:
-        # TODO: see if np.put() or other alternatives works faster (requires flatten array)
-        sal_maps_tmp = sal_maps
-        sal_maps = np.zeros((num_classes, input_size[0], input_size[1]))
-        for i, sal in enumerate(sal_maps_tmp):
-            sal_maps[target_classes[i]] = sal
-        return sal_maps
+        return {target_class: saliency_map for target_class, saliency_map in zip(target_classes, saliency_maps)}
 
     @staticmethod
     def _generate_mask(input_size: Tuple[int, int], num_cells: int, prob: float, rand_generator) -> np.ndarray: