+++ tests/unit/explanation/test_explainer.py
@@ -10,10 +10,11 @@ import pytest
 from pytest_mock import MockerFixture
 
 from openvino_xai.api.api import insert_xai
-from openvino_xai.common.parameters import Task
+from openvino_xai.common.parameters import Method, Task
 from openvino_xai.common.utils import retrieve_otx_model
 from openvino_xai.explainer.explainer import Explainer, ExplainMode
 from openvino_xai.explainer.utils import get_postprocess_fn, get_preprocess_fn
+from openvino_xai.methods.black_box.base import Preset
 from tests.unit.explanation.test_explanation_utils import VOC_NAMES
 
 MODEL_NAME = "mlc_mobilenetv3_large_voc"
@@ -39,15 +40,8 @@ class TestExplainer:
             ExplainMode.AUTO,
         ],
     )
-    @pytest.mark.parametrize(
-        "explain_all_classes",
-        [
-            True,
-            False,
-        ],
-    )
     @pytest.mark.parametrize("with_xai_originally", [True, False])
-    def test_explainer(self, explain_mode, explain_all_classes, with_xai_originally):
+    def test_explainer(self, explain_mode, with_xai_originally):
         retrieve_otx_model(self.data_dir, MODEL_NAME)
         model_path = self.data_dir / "otx_models" / (MODEL_NAME + ".xml")
         model = ov.Core().read_model(model_path)
@@ -66,19 +60,96 @@ class TestExplainer:
             explain_mode=explain_mode,
         )
 
+        explanation = explainer(
+            self.image,
+            targets=[11, 14],
+            label_names=VOC_NAMES,  # optional
+            preset=Preset.SPEED,
+        )
+
+        assert len(explanation.saliency_map) == 2
+
+    @pytest.mark.parametrize(
+        "explain_all_classes",
+        [
+            True,
+            False,
+        ],
+    )
+    def test_explainer_all_classes(self, explain_all_classes):
+        retrieve_otx_model(self.data_dir, MODEL_NAME)
+        model_path = self.data_dir / "otx_models" / (MODEL_NAME + ".xml")
+        model = ov.Core().read_model(model_path)
+
+        explainer = Explainer(
+            model=model,
+            task=Task.CLASSIFICATION,
+            preprocess_fn=self.preprocess_fn,
+            postprocess_fn=get_postprocess_fn(),
+        )
+
         if explain_all_classes:
             explanation = explainer(
                 self.image,
                 targets=-1,
                 label_names=VOC_NAMES,  # optional
-                num_masks=10,
             )
         else:
             explanation = explainer(
                 self.image,
                 targets=[11, 14],
                 label_names=VOC_NAMES,  # optional
-                num_masks=10,
+            )
+
+        if explain_all_classes:
+            assert len(explanation.saliency_map) == 20
+        else:
+            assert len(explanation.saliency_map) == 2
+
+    @pytest.mark.parametrize(
+        "explain_all_classes",
+        [
+            True,
+            False,
+        ],
+    )
+    @pytest.mark.parametrize(
+        "explain_method",
+        [
+            Method.AISE,
+            Method.RISE,
+        ],
+    )
+    def test_explainer_all_classes_black_box(self, explain_all_classes, explain_method):
+        retrieve_otx_model(self.data_dir, MODEL_NAME)
+        model_path = self.data_dir / "otx_models" / (MODEL_NAME + ".xml")
+        model = ov.Core().read_model(model_path)
+
+        explainer = Explainer(
+            model=model,
+            task=Task.CLASSIFICATION,
+            preprocess_fn=self.preprocess_fn,
+            postprocess_fn=get_postprocess_fn(),
+            explain_mode=ExplainMode.BLACKBOX,
+            explain_method=explain_method,
+        )
+
+        if explain_method == Method.AISE:
+            kwargs = {"num_iterations_per_kernel": 2}
+        else:
+            kwargs = {"num_masks": 2}
+
+        if explain_all_classes:
+            explanation = explainer(
+                self.image,
+                targets=-1,
+                **kwargs,
+            )
+        else:
+            explanation = explainer(
+                self.image,
+                targets=[11, 14],
+                **kwargs,
             )
 
         if explain_all_classes:
@@ -113,7 +184,7 @@ class TestExplainer:
         )
 
         processed_data = self.preprocess_fn(self.image)
-        explanation = explainer(processed_data, targets=[11, 14], num_masks=10)
+        explanation = explainer(processed_data, targets=[11, 14], preset=Preset.SPEED)
 
         assert len(explanation.saliency_map) == 2
 
@@ -140,7 +211,7 @@ class TestExplainer:
             explain_mode=ExplainMode.AUTO,
             target_layer="some_wrong_name",
         )
-        explanation = explainer(self.image, targets=-1, num_masks=10)
+        explanation = explainer(self.image, targets=-1, num_iterations_per_kernel=2)
         assert len(explanation.saliency_map) == 20
 
     def test_two_explainers_with_same_model(self):