+++ tests/unit/explanation/test_explanation_utils.py
@@ -4,12 +4,8 @@
 import numpy as np
 import pytest
 
-from openvino_xai.explainer.utils import (
-    ActivationType,
-    get_explain_target_indices,
-    get_score,
-    is_bhwc_layout,
-)
+from openvino_xai.common.utils import is_bhwc_layout
+from openvino_xai.explainer.utils import ActivationType, get_score, get_target_indices
 
 VOC_NAMES = [
     "aeroplane",
@@ -38,25 +34,25 @@ LABELS_INT = [1, 4]
 LABELS_STR = ["bicycle", "bottle"]
 
 
-def test_get_explain_target_indices_int():
-    explain_target_indices = get_explain_target_indices(LABELS_INT, VOC_NAMES)
-    assert np.all(explain_target_indices == LABELS_INT)
+def test_get_target_indices_int():
+    target_indices = get_target_indices(LABELS_INT, VOC_NAMES)
+    assert np.all(target_indices == LABELS_INT)
 
 
-def test_get_explain_target_indices_int_wo_names():
-    explain_target_indices = get_explain_target_indices(LABELS_INT)
-    assert np.all(explain_target_indices == LABELS_INT)
+def test_get_target_indices_int_wo_names():
+    target_indices = get_target_indices(LABELS_INT)
+    assert np.all(target_indices == LABELS_INT)
 
 
-def test_get_explain_target_indices_str():
-    explain_target_indices = get_explain_target_indices(LABELS_STR, VOC_NAMES)
-    assert explain_target_indices == [1, 4]
+def test_get_target_indices_str():
+    target_indices = get_target_indices(LABELS_STR, VOC_NAMES)
+    assert target_indices == [1, 4]
 
 
-def test_get_explain_target_indices_str_spelling():
+def test_get_target_indices_str_spelling():
     LABELS_STR[0] = "bicycle_"
     with pytest.raises(Exception) as exc_info:
-        _ = get_explain_target_indices(LABELS_STR, VOC_NAMES)
+        _ = get_target_indices(LABELS_STR, VOC_NAMES)
     assert str(exc_info.value) == "No all label names found in label_names. Check spelling."